backend/requirements.txt
# Python 3.10+
# NumPy 2.x breaks many PyTorch wheels; opencv-python 4.12+ often requires NumPy 2.
# This combo keeps YOLO + OpenCV + MediaPipe on one machine without forcing torch nightly.
numpy>=1.26.2,<2
opencv-python>=4.8.1,<4.12.0
mediapipe>=0.10.14,<0.11
ultralytics>=8.0.200
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
